Fix serializer fields

+++ backend/phones/models.py
@@ -3,6 +3,6 @@ from django.db import models
 
 class Phone(models.Model):
     value = models.CharField(max_length=50)
-    monthy_price = models.IntegerField(default=1)
+    monthy_price = models.FloatField(default=1.0)
     currency = models.CharField(max_length=10)
-    setup_price = models.IntegerField(default=1)
+    setup_price = models.FloatField(default=1.0)
